--- src/main/java/inu/codin/codin/domain/elasticsearch/event/LectureSavedEvent.java
@@ -1,11 +1,0 @@
-package inu.codin.codin.domain.elasticsearch.event;
-
-import inu.codin.codin.domain.lecture.entity.Lecture;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class LectureSavedEvent {
-    private Lecture lecture;
-}
