+++ src/main/java/inu/codin/codin/domain/lecture/event/LectureSavedEvent.java
@@ -1,0 +1,6 @@
+package inu.codin.codin.domain.lecture.event;
+
+import inu.codin.codin.domain.lecture.entity.Lecture;
+
+public record LectureSavedEvent(Lecture lecture) {
+}
