--- src/main/java/inu/codin/codin/domain/lecture/repository/LectureSearchRepositoryCustom.java
@@ -1,17 +1,0 @@
-package inu.codin.codin.domain.lecture.repository;
-
-import inu.codin.codin.domain.elasticsearch.document.LectureDocument;
-import inu.codin.codin.domain.lecture.entity.Lecture;
-import inu.codin.codin.domain.lecture.entity.Semester;
-import inu.codin.codin.domain.lecture.entity.SortingOption;
-import inu.codin.codin.global.common.entity.Department;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-
-import java.util.List;
-
-public interface LectureSearchRepositoryCustom {
-    Page<LectureDocument> searchLectureList(String keyword, Department department, SortingOption sortingOption, List<String> likeIdList, Pageable pageable, Boolean like);
-
-    List<Lecture> searchLecturesAtReview(Department department, Integer grade, Semester semester);
-}
